+++ src/components/input-panel/compiled-spec-display/index.tsx
@@ -2,7 +2,7 @@ import stringify from 'json-stringify-pretty-compact';
 import * as React from 'react';
 import MonacoEditor from 'react-monaco-editor';
 import { connect } from 'react-redux';
-import ConfigEditor from '../../theme-editor/index';
+import ConfigEditor from '../../config-editor/index';
 
 import { Mode, SIDEPANE } from '../../../constants';
 import { State } from '../../../constants/default-state';
